test(log-store): split artifact repository coverage
